Improve HTML tag handling in the parser
changes include:
Correctly identifying block tags and pushing them onto the stack. - Michailkr
Properly closing matching open and close tags by removing them from the stack. - MIchailkr
Flushing the output buffer to an HTML file when the buffer is full, and maintaining the stack state across buffer flushes. - Lyapin Ermak
Handling the case where a tag is not a block tag and there is not enough space in the buffer to write it. - Lyapin Ermak

Co-Authored-By: Lyapin Ermak <166688928+xx-0-razzbivatel-4ereepov-0-xx@users.noreply.github.com>