Improve HTML tag handling in the parser
changes include:
Correctly identifying block tags and pushing them onto the stack. - Michailkr
Properly closing matching open and close tags by removing them from the stack. - MIchailkr
Flushing the output buffer to an HTML file when the buffer is full, and maintaining the stack state across buffer flushes. - RaZZbivatel-4ErEePOV
Handling the case where a tag is not a block tag and there is not enough space in the buffer to write it. - RaZZbivatel-4ErEePOV

Co-Authored-By: xX-0-RaZZbivatel-4ErEePOV-0-Xx <166688928+xx-0-razzbivatel-4ereepov-0-xx@users.noreply.github.com>