readme and finishing library
I created a README
and introduced new features.
The first function checks the character read from the data stream to determine if it is the opening character of the XML tag ("<"). If the character read from the stream is "<", then the is_tag_open variable is set to 1, which means that the tag is open. Then the symbol is written to the tag array at the index_tag index, after which the index is increased by one (index_tag++) so that at the next character it is stored at the next position in the tag array.
The second function checks the condition when the tag is not blocky and there is enough buffer space to write it.

If there is enough buffer space for the tag (max_len - len_output) greater than the tag length (strlen(tag)), and also taking into account the length of the stack contents (len_stack), then the following is performed:
- The string contained in the tag variable is copied to the output buffer, starting at the len_output position.
- len_output increases by the length of the tag string (strlen(tag)) to indicate a new position in the buffer after writing the tag.
- index_output increases by the length of the tag string (strlen(tag)), presumably this is a variable for tracking the position in the buffer.
- The tag array is reset to zeros (memset(tag, '\0', 500)), to prepare it for the next entry of the new tag.
- The index for writing characters to the tag array (index_tag) is reset to 0.

This section of code allows you to write tags to the buffer if they are not blocky and there is enough space to place them, given the length of the existing content